runtime (scheduler): generalize scheduler beyond timers
