+++ src/runtime/scheduler_any.go
@@ -19,6 +19,7 @@ func run() {
 		initAll()
 		postinit()
 		callMain()
+		schedulerDone = true
 	}()
 	scheduler()
 }
